Cring boomer

+++ Afg4Urlaubsfahrt/src/Urlaubsfahrt/Car.cs
@@ -5,29 +5,29 @@
         /// <summary>
         /// The fuel usage of the car. measured in Litres / Kilometres.
         /// </summary>
-        public readonly double FuelUsage;
+        public readonly decimal FuelUsage;
 
         /// <summary>
         /// The capacity of the cars tank, measured in Litres.
         /// </summary>
-        public readonly double TankCapacity;
+        public readonly decimal TankCapacity;
 
         /// <summary>
         /// The distance the car can travel on one tank without refueling, measured in Metres.
         /// </summary>
-        public readonly double TankDistance;
+        public readonly decimal TankDistance;
 
         /// <summary>
         /// The amount of fuel the car starts off with, measured in Litres.
         /// </summary>
-        public readonly double StartingFuelAmount;
+        public readonly decimal StartingFuelAmount;
 
         /// <summary>1
         /// The distance the car can travel on with the fuel it starded with without refueling, measured in Metres.
         /// </summary>
-        public readonly double StartingFuelDistance;
+        public readonly decimal StartingFuelDistance;
 
-        public Car(double fuelUsage, double tankCapacity, double startingFuelAmount)
+        public Car(decimal fuelUsage, decimal tankCapacity, decimal startingFuelAmount)
         {
             FuelUsage = fuelUsage;
             TankCapacity = tankCapacity;
@@ -41,6 +41,6 @@
         /// </summary>
         /// <param name="station">The gas station to get the price for.</param>
         /// <returns>The price in € of driving a kilometer with gas from this station.</returns>
-        public readonly double GetPriceForDistanceAt(GasStation station) => station.Price * FuelUsage;
+        public readonly decimal GetPriceForDistanceAt(GasStation station) => station.Price * FuelUsage;
     }
 }
